backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
qdrant-client==1.7.0
sentence-transformers==2.3.1
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
google-generative-ai==0.3.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0